CT-4498
(cherry picked from commit eac9da082aa1b757013b37b587afd523c31e5b1b)

+++ i21Database/dbo/Stored Procedures/uspCTCreateInvoiceDetail.sql
@@ -92,7 +92,7 @@ BEGIN TRY
 				,@ItemDestinationGradeId				INT             = NULL
 				,@ItemDestinationWeightId				INT             = NULL
 				,@intScaleUOMId							INT  
-				,@ItemPrice								INT  
+				,@ItemPrice								NUMERIC(18,8)  
 				,@InvoiceId								INT
 
 		SELECT	TOP 1
